requirements.txt
python-dotenv>=1.0.1
requests>=2.32.3
pandas>=2.2.2
duckdb>=1.0.0
pyarrow>=17.0.0
tenacity>=8.3.0
